refactored logic into services
