+++ Webapp/sources/blueprints/save_reaction/routes.py
@@ -2,16 +2,11 @@
 This module receives a reaction name from the reaction
 table and saves it in the reaction database
 """
-import base64
-import io
-import os
-import uuid
+
 from datetime import datetime
 
-import magic
 import pytz
-from azure.core.exceptions import ResourceExistsError
-from flask import Response, abort, current_app, json, jsonify, request
+from flask import Response, abort, json, jsonify, request
 from flask_login import current_user, login_required
 from sources import models, services
 from sources.auxiliary import (
@@ -22,7 +17,6 @@ from sources.auxiliary import (
 )
 from sources.extensions import db
 from sqlalchemy import func
-from werkzeug.utils import secure_filename
 
 from . import save_reaction_bp
 
@@ -161,7 +155,7 @@ def new_reaction() -> Response:
 def autosave() -> Response:
     """autosave when a field changes in the reaction page"""
     reaction_description = str(request.form["reactionDescription"])
-    reaction = get_current_reaction()
+    reaction = services.reaction.get_current()
     reaction_name = reaction.name
 
     services.auth.edit_reaction(reaction)
@@ -401,7 +395,7 @@ def autosave() -> Response:
 def autosave_sketcher() -> Response:
     """Autosave function for saving changes to the sketcher only. Only used before reaction table is generated."""
 
-    reaction = get_current_reaction()
+    reaction = services.reaction.get_current()
     services.auth.edit_reaction(reaction)
 
     current_time = datetime.now(pytz.timezone("Europe/London")).replace(tzinfo=None)
@@ -451,8 +445,8 @@ def check_reaction_name() -> Response:
 @login_required
 def upload_experiment_files():
     """Takes a list of files, and saves upon successful validation. Url added to database, file saved to azure blob"""
-    authenticate_user(permission_level="edit", request_method="POST")
-    new_upload = UploadExperimentDataFiles(request)
+    services.auth.reaction(permission_level="edit", request_method="POST")
+    new_upload = services.file_attachments.UploadExperimentDataFiles(request)
     new_upload.validate_files()
     new_upload.save_validated_files()
     return jsonify({"uploaded_files": new_upload.uploaded_files})
@@ -464,7 +458,7 @@ def view_reaction_attachment() -> Response:
     """
     Authenticate user has permission to view, then use the uuid to find the file on azure and then return the file.
     """
-    authenticate_user(permission_level="view_only", request_method="GET")
+    services.auth.reaction(permission_level="view_only", request_method="GET")
     file_uuid = request.args.get("uuid")
     # get the blob file and send the filestream, display name and mimetype to the frontend to be displayed.
     blob_client = services.file_attachments.get_blob(file_uuid)
@@ -488,7 +482,7 @@ def view_reaction_attachment() -> Response:
 @login_required
 def download_experiment_files():
     """Take a file and return as attachment to the user"""
-    authenticate_user(permission_level="view_only", request_method="POST")
+    services.auth.reaction(permission_level="view_only", request_method="POST")
     blob_client = services.file_attachments.get_blob()
     file_uuid = request.form["uuid"]
     file_attachment = services.file_attachments.download_from_blob_client(
@@ -506,7 +500,7 @@ def download_experiment_files():
 @login_required
 def delete_reaction_attachment():
     """Delete file attached to reaction"""
-    authenticate_user(permission_level="edit", request_method="DELETE")
+    services.auth.reaction(permission_level="edit", request_method="DELETE")
     # connect to blob
     blob_client = services.file_attachments.get_blob()
     # delete blob
@@ -520,169 +514,3 @@ def delete_reaction_attachment():
     db.session.delete(file_object)
     db.session.commit()
     return "success"
-
-
-def authenticate_user(permission_level: str, request_method: str):
-    """
-    Authenticates user to either view or edit the reaction.
-    permission_level: Takes value of 'edit' or 'view_only'
-    request_method: Value of 'GET' changes behaviour, other strings all have same behaviour (e.g. POST, DELETE)
-    """
-    if permission_level == "view_only":
-        authenticate_user_to_view_files(request_method)
-    if permission_level == "edit":
-        reaction = get_current_reaction()
-        services.auth.edit_reaction(reaction, file_attachment=True)
-
-
-def authenticate_user_to_view_files(request_method):
-    """Authenticates user as a workbook member or aborts. Gets the workgroup_name, workbook_name, and workbook."""
-    if request_method == "GET":
-        workgroup_name = request.args.get("workgroup")
-        workbook_name = request.args.get("workbook")
-    else:
-        workgroup_name = request.form["workgroup"]
-        workbook_name = request.form["workbook"]
-    # validate user belongs to the workbook
-    workbook = services.workbook.get_workbook_from_group_book_name_combination(
-        workgroup_name, workbook_name
-    )
-    abort_if_user_not_in_workbook(workgroup_name, workbook_name, workbook)
-
-
-class UploadExperimentDataFiles:
-    def __init__(self, ajax_request):
-        self.files_to_upload = ajax_request.files
-        self.institution = db.session.query(models.Institution).first()
-        self.workgroup = ajax_request.form["workgroup"]
-        self.workbook = ajax_request.form["workbook"]
-        self.reaction_id = ajax_request.form["reactionID"]
-        self.validated_files = []
-        self.container_client = None
-        self.reaction = get_current_reaction()
-        self.uploaded_files = []
-        self.file_hash = ""
-
-    def validate_number_of_attachments(self):
-        number_of_attachments = len(self.reaction.file_attachments)
-        if number_of_attachments >= 10:
-            abort(401)
-
-    def validate_files(self):
-        for file in self.files_to_upload.values():
-            validation_result = self.file_security_validation(file)
-            if validation_result == "success":
-                self.validated_files.append(file)
-            else:
-                abort(401)
-
-    def save_validated_files(self):
-        for file in self.validated_files:
-            # file name must be unique - needs workgroup/book/reaction_id/filename. Check for uniqueness of filename.
-            filename = str(uuid.uuid4())
-            self.save_blob(file, filename)
-            print("blob uploaded")
-            # save file details to database, and shorten filename if too long.
-            name, extension = os.path.splitext(file.filename)
-            if len(name) > 21:
-                file.filename = name[0:10] + "..." + name[-10:-1] + extension
-            self.save_file_details_to_database(file, filename)
-            self.uploaded_files.append({"uuid": filename, "name": file.filename})
-
-    def save_file_details_to_database(self, file, filename):
-        current_time = datetime.now(pytz.timezone("Europe/London")).replace(tzinfo=None)
-        storage_name = self.container_client.account_name
-        container_name = self.container_client.container_name
-        file_size = os.fstat(file.fileno()).st_size
-        file_ext = os.path.splitext(file.filename)[1]
-        file_details = {
-            "mimetype": file.mimetype,
-            "size": file_size,
-            "extension": file_ext,
-        }
-        models.ReactionDataFile.create(
-            reaction=self.reaction.id,
-            storage_name=storage_name,
-            container_name=container_name,
-            uuid=filename,
-            display_name=file.filename,
-            time_of_upload=current_time,
-            file_details=file_details,
-            sha256_checksum=self.file_hash,
-        )
-
-    def save_blob(self, file, filename):
-        # connect to storage account
-        blob_service_client = services.file_attachments.connect_to_azure_blob_service()
-        # connect to container
-        container_name = current_app.config["STORAGE_CONTAINER"]
-        try:
-            self.container_client = blob_service_client.create_container(container_name)
-        except ResourceExistsError:
-            self.container_client = blob_service_client.get_container_client(
-                container_name
-            )
-
-        blob_client = blob_service_client.get_blob_client(
-            container=container_name, blob=filename
-        )
-        print("uploading blob")
-        self.file_hash = services.file_attachments.sha256_from_file_contents(
-            file.stream.seek(0).stream.read()
-        )
-        if not self.file_hash:
-            print("Failed to generate hash")
-            abort(500)
-        # Upload the blob data - default blob type is BlockBlob
-        file.stream.seek(0)
-        upload2 = io.BytesIO(file.stream.read())
-        blob_client.upload_blob(upload2, blob_type="BlockBlob")
-        # confirm upload
-        if not blob_client.exists():
-            print(f"blob {filename} upload failed")
-            abort(401)
-
-    @staticmethod
-    def file_security_validation(file):
-        """Validates the filesize, name, extension, and mimetype"""
-        filename = file.filename
-        secure_filename(filename)
-        if filename != "":
-            # size under 1 mb
-            filesize = os.fstat(file.fileno()).st_size
-            if filesize > 1000000:
-                print("file too large")
-                return "failure"
-            # acceptable file extension
-            file_ext = os.path.splitext(filename)[1]
-            if file_ext not in current_app.config["UPLOAD_EXTENSIONS"]:
-                print("incorrect extension")
-                return "failure"
-            # acceptable mimetype
-            mime = magic.Magic(mime=True)
-            mime_type = mime.from_buffer(file.read(2048))
-            if mime_type not in current_app.config["UPLOAD_MIME_TYPES"]:
-                print("incorrect mimetype")
-                return "failure"
-        else:
-            return "failure"
-        return "success"
-
-
-def get_current_reaction():
-    reaction_id = str(request.form["reactionID"])
-    workgroup_name = str(request.form["workgroup"])
-    workbook_name = str(request.form["workbook"])
-    workbook = services.workbook.get_workbook_from_group_book_name_combination(
-        workgroup_name, workbook_name
-    )
-    abort_if_user_not_in_workbook(workgroup_name, workbook_name, workbook)
-    return (
-        db.session.query(models.Reaction)
-        .filter(models.Reaction.reaction_id == reaction_id)
-        .join(models.WorkBook)
-        .filter(models.WorkBook.id == workbook.id)
-        .join(models.WorkGroup)
-        .filter(models.WorkGroup.name == workgroup_name)
-        .first()
-    )
